Basic tileset and item rendering pipeline in place. Lots more to do.

+++ src/components/renderable.hpp
@@ -5,13 +5,13 @@
 using namespace rltk;
 
 struct renderable_t {
-	uint8_t glyph;
+	uint16_t glyph;
 	color_t foreground;
 	color_t background;
 	std::string xml_identity = "renderable_t";
 
 	renderable_t() {}
-	renderable_t(const uint8_t ch, const color_t fg, const color_t bg) : glyph(ch), foreground(fg), background(bg) {}
+	renderable_t(const uint16_t ch, const color_t fg, const color_t bg) : glyph(ch), foreground(fg), background(bg) {}
 
 	void to_xml(xml_node * c);
 	void from_xml(xml_node * c);
